only upload if there isnt one

+++ cli/driver-doc-to-s3.php
@@ -13,11 +13,17 @@ $path = '/home/cockpit.la/www/upload/drivers-doc/';
 
 S3::setAuth($key, $secret);
 
+$contents = S3::getBucket($bucket);
+
 $dir = new DirectoryIterator($path);
 foreach ($dir as $fileinfo) {
 	if (!$fileinfo->isDot()) {
 		$file = $fileinfo->getFilename();
 		echo "Uploading $file...";
+		if (array_key_exists($file, $contents)) {
+			echo "skipped\n";
+			continue;
+		}
 		$r = S3::putObject(S3::inputFile($path.$file, false), $bucket, $file, S3::ACL_PRIVATE);
 		echo ($r ? 'success' : 'FAILED')."\n";
 		exit;
